added support for reading metadata tags from ogg vorbis files.  you can view them in wxPlayer.
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@571 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ src/device_mixer.cpp
@@ -91,7 +91,7 @@ namespace audiere {
         } else if (mixed > 32767) {
           mixed = 32767;
         }
-        *out++ = mixed;
+        *out++ = (s16)mixed;
       }
 
       left -= to_mix;
